Split typechecker semantic validation
