+++ docs/PHASE_PLAN.md
@@ -150,6 +150,9 @@ checked-in docs, tests, and commits only.
 - Typechecker resolver-backed declaration collection orchestration now lives in
   a dedicated helper module, keeping resolver handoff passes out of the broad
   typechecker root.
+- Typechecker semantic validation now lives in a dedicated helper module,
+  keeping behavior association validation and struct-field default checks out
+  of the broad typechecker root.
 - Resolver validation replay now collects expected declaration symbols,
   expected local symbols, import-validation state, and behavior-association
   replay tasks in one declaration pass before checking resolver-owned extras,
