+++ src/typechecker/semantic_validation.rs
@@ -1,0 +1,354 @@
+use super::*;
+
+impl TypeChecker {
+    pub(super) fn validate_collected_declaration_semantics(
+        &mut self,
+        decls: &[Declaration],
+        symbols: Option<&SymbolTable>,
+    ) {
+        if self.resolver_backed_collection {
+            let tasks = Self::collect_resolver_semantic_validation_tasks(decls);
+            self.validate_behavior_association_tasks(&tasks, symbols);
+            self.validate_resolver_type_reference_tasks(&tasks, symbols);
+            self.validate_resolver_struct_field_default_tasks(&tasks, symbols);
+            return;
+        }
+
+        let tasks = Self::collect_ast_declaration_validation_tasks(decls);
+        self.validate_behavior_association_tasks(&tasks.behavior_associations, symbols);
+        self.validate_ast_type_reference_tasks(&tasks.type_references);
+        self.validate_ast_struct_field_default_tasks(&tasks.struct_field_defaults);
+    }
+
+    pub(super) fn collect_ast_declaration_validation_tasks(
+        decls: &[Declaration],
+    ) -> AstDeclarationValidationTasks<'_> {
+        let mut tasks = AstDeclarationValidationTasks::default();
+        for decl in decls {
+            Self::push_behavior_extends_replay_task(decl, &mut tasks.behavior_associations.extends);
+            Self::push_behavior_impl_block_declaration_task(
+                decl,
+                &mut tasks.behavior_associations.impls,
+            );
+            Self::push_behavior_requires_replay_task(
+                decl,
+                &mut tasks.behavior_associations.requires,
+            );
+            Self::push_ast_type_reference_validation_task(decl, &mut tasks.type_references);
+            Self::push_ast_struct_field_default_validation_task(
+                decl,
+                &mut tasks.struct_field_defaults,
+            );
+        }
+        tasks
+    }
+
+    #[cfg(test)]
+    pub(super) fn collect_behavior_association_validation_tasks(
+        decls: &[Declaration],
+    ) -> BehaviorAssociationValidationTasks<'_> {
+        let mut tasks = BehaviorAssociationValidationTasks::default();
+        for decl in decls {
+            Self::push_behavior_extends_replay_task(decl, &mut tasks.extends);
+            Self::push_behavior_impl_block_declaration_task(decl, &mut tasks.impls);
+            Self::push_behavior_requires_replay_task(decl, &mut tasks.requires);
+        }
+        tasks
+    }
+
+    fn collect_resolver_semantic_validation_tasks(
+        decls: &[Declaration],
+    ) -> ResolverDeclarationMetadataTasks<'_> {
+        Self::collect_resolver_declaration_metadata_tasks(decls)
+    }
+
+    pub(super) fn validate_behavior_association_tasks<'a>(
+        &mut self,
+        tasks: &impl BehaviorAssociationValidationTaskSource<'a>,
+        symbols: Option<&SymbolTable>,
+    ) {
+        let tasks = tasks.behavior_association_tasks();
+        self.validate_behavior_impl_tasks(tasks, symbols);
+        self.validate_behavior_requires_tasks(tasks, symbols);
+    }
+
+    fn validate_behavior_impl_tasks(
+        &mut self,
+        tasks: &BehaviorAssociationValidationTasks<'_>,
+        symbols: Option<&SymbolTable>,
+    ) {
+        for task in &tasks.impls {
+            self.validate_collected_behavior_impl_declaration(
+                symbols,
+                task.ast_type_name,
+                task.behavior,
+                task.behavior_type_args,
+                task.methods,
+                task.span,
+            );
+        }
+    }
+
+    pub(super) fn push_behavior_requires_replay_task<'a>(
+        decl: &'a Declaration,
+        tasks: &mut Vec<BehaviorRequiresValidationTask<'a>>,
+    ) -> bool {
+        if let Declaration::Requires {
+            type_name,
+            behavior,
+            behavior_type_args,
+            span,
+        } = decl
+        {
+            tasks.push(BehaviorRequiresValidationTask {
+                type_name,
+                behavior,
+                behavior_type_args,
+                span: *span,
+            });
+            true
+        } else {
+            false
+        }
+    }
+
+    fn validate_behavior_requires_tasks(
+        &mut self,
+        tasks: &BehaviorAssociationValidationTasks<'_>,
+        symbols: Option<&SymbolTable>,
+    ) {
+        for task in &tasks.requires {
+            self.validate_collected_behavior_requires_declaration(
+                symbols,
+                task.type_name,
+                task.behavior,
+                task.behavior_type_args,
+                task.span,
+            );
+        }
+    }
+
+    fn validate_collected_behavior_impl_declaration(
+        &mut self,
+        symbols: Option<&SymbolTable>,
+        type_name: &str,
+        behavior: &str,
+        behavior_type_args: &[AstType],
+        methods: &[Declaration],
+        span: Span,
+    ) {
+        let restored_type_name = symbols
+            .map(|symbols| {
+                self.resolver_impl_type_name_for(
+                    symbols,
+                    type_name,
+                    methods,
+                    Some((behavior, behavior_type_args)),
+                )
+            })
+            .unwrap_or_else(|| type_name.to_string());
+        self.check_behavior_impl(
+            &restored_type_name,
+            behavior,
+            behavior_type_args,
+            methods,
+            span,
+            symbols,
+        );
+    }
+
+    fn validate_collected_behavior_requires_declaration(
+        &mut self,
+        symbols: Option<&SymbolTable>,
+        type_name: &str,
+        behavior: &str,
+        behavior_type_args: &[AstType],
+        span: Span,
+    ) {
+        let type_name = symbols
+            .map(|symbols| {
+                self.resolver_required_type_name_for(
+                    symbols,
+                    type_name,
+                    behavior,
+                    behavior_type_args,
+                )
+            })
+            .unwrap_or_else(|| type_name.to_string());
+        self.check_behavior_requires(&type_name, behavior, behavior_type_args, span);
+    }
+
+    #[cfg(test)]
+    pub(super) fn validate_struct_field_defaults(
+        &mut self,
+        decls: &[Declaration],
+        symbols: Option<&SymbolTable>,
+    ) {
+        if self.resolver_backed_collection {
+            let tasks = Self::collect_resolver_declaration_metadata_tasks(decls);
+            self.validate_resolver_struct_field_default_tasks(&tasks, symbols);
+            return;
+        }
+
+        let tasks = Self::collect_ast_struct_field_default_validation_tasks(decls);
+        self.validate_ast_struct_field_default_tasks(&tasks);
+    }
+
+    #[cfg(test)]
+    pub(super) fn collect_ast_struct_field_default_validation_tasks(
+        decls: &[Declaration],
+    ) -> Vec<AstStructFieldDefaultValidationTask<'_>> {
+        let mut tasks = Vec::new();
+        for decl in decls {
+            Self::push_ast_struct_field_default_validation_task(decl, &mut tasks);
+        }
+        tasks
+    }
+
+    fn push_ast_struct_field_default_validation_task<'a>(
+        decl: &'a Declaration,
+        tasks: &mut Vec<AstStructFieldDefaultValidationTask<'a>>,
+    ) {
+        if let Declaration::Struct {
+            type_params,
+            fields,
+            ..
+        } = decl
+        {
+            tasks.push(AstStructFieldDefaultValidationTask {
+                type_params,
+                fields,
+            });
+        }
+    }
+
+    fn validate_ast_struct_field_default_tasks(
+        &mut self,
+        tasks: &[AstStructFieldDefaultValidationTask<'_>],
+    ) {
+        for task in tasks {
+            self.validate_ast_struct_field_defaults(!task.type_params.is_empty(), task.fields);
+        }
+    }
+
+    fn validate_resolver_struct_field_defaults(
+        &mut self,
+        symbols: Option<&SymbolTable>,
+        name: &str,
+        span: Span,
+    ) {
+        let restored_name = Self::validation_symbol_name(symbols, Namespace::Type, name, span);
+        let Some(info) = self.structs.get(&restored_name).cloned() else {
+            return;
+        };
+        if !info.type_params.is_empty() {
+            return;
+        }
+        for (field_name, expected) in &info.fields {
+            if let Some(default) = info.field_defaults.get(field_name) {
+                self.validate_struct_field_default(field_name, expected, default);
+            }
+        }
+    }
+
+    pub(super) fn validate_resolver_struct_field_default_tasks(
+        &mut self,
+        tasks: &ResolverDeclarationMetadataTasks<'_>,
+        symbols: Option<&SymbolTable>,
+    ) {
+        for task in &tasks.types {
+            if let ResolverTypeDeclarationMetadataTask::Struct { name, span, .. } = task {
+                self.validate_resolver_struct_field_defaults(symbols, name, *span);
+            }
+        }
+    }
+
+    fn validate_ast_struct_field_defaults(
+        &mut self,
+        has_type_params: bool,
+        fields: &[StructField],
+    ) {
+        if has_type_params {
+            return;
+        }
+        for field in fields {
+            if let Some(default) = &field.default {
+                self.validate_struct_field_default(&field.name, &field.ty, default);
+            }
+        }
+    }
+
+    fn validate_struct_field_default(
+        &mut self,
+        field_name: &str,
+        expected: &AstType,
+        default: &Expression,
+    ) {
+        let expected = self.resolve_type(expected);
+        self.push_scope();
+        let actual = self.check_expr(default);
+        self.pop_scope();
+
+        let Ok(actual) = actual else {
+            self.diagnostics.push(actual.expect_err("checked error"));
+            return;
+        };
+        let actual_ty = if (expected.is_integer()
+            && matches!(actual.kind, TypedExprKind::IntLiteral(_)))
+            || (expected.is_float() && matches!(actual.kind, TypedExprKind::FloatLiteral(_)))
+        {
+            expected.clone()
+        } else {
+            actual.ty.clone()
+        };
+
+        if !self.types_compatible(&expected, &actual_ty) {
+            self.diagnostics.push(Diagnostic::error(
+                "E3073",
+                format!(
+                    "field `{}` default expects `{}`, found `{}`",
+                    field_name,
+                    expected.display_name(),
+                    actual.ty.display_name()
+                ),
+                actual.span,
+            ));
+        }
+    }
+
+    pub(super) fn validate_collected_behavior_extends_semantics(&mut self) {
+        let behavior_extends: Vec<(String, Vec<BehaviorParentRef>, Span)> = self
+            .behavior_extends
+            .iter()
+            .map(|(behavior, parents)| {
+                (
+                    behavior.clone(),
+                    parents.clone(),
+                    self.behavior_extends_spans
+                        .get(behavior)
+                        .copied()
+                        .unwrap_or_else(Span::dummy),
+                )
+            })
+            .collect();
+
+        for (behavior, parents, span) in behavior_extends {
+            let scoped_type_params: HashSet<String> = self
+                .behaviors
+                .get(&behavior)
+                .map(|info| info.type_params.iter().cloned().collect())
+                .unwrap_or_default();
+            for parent in parents {
+                self.behavior_type_arg_substitutions(
+                    &parent.behavior,
+                    &parent.type_args,
+                    &scoped_type_params,
+                    span,
+                );
+            }
+        }
+
+        self.validate_behavior_extends_cycles();
+        self.validate_behavior_method_coherence();
+    }
+}
